+++ .travis.yml
@@ -1,0 +1,15 @@
+sudo: false
+
+language: go
+env:
+  - GO111MODULE = on
+
+go:
+  - 1.11.x
+  - master
+
+before_script:
+  - go mod download
+
+script:
+  - go test -v -race ./...  # Run all the tests with the race detector enabled
