+++ go.mod
@@ -4,25 +4,20 @@ require (
 	github.com/PuerkitoBio/purell v1.1.1 // indirect
 	github.com/alecthomas/template v0.0.0-20160405071501-a0175ee3bccc
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/gin-contrib/sse v0.0.0-20190301062529-5545eab6dad3 // indirect
 	github.com/go-chi/chi v4.0.2+incompatible // indirect
 	github.com/go-openapi/jsonpointer v0.18.0 // indirect
 	github.com/go-openapi/swag v0.18.0 // indirect
 	github.com/go-sql-driver/mysql v1.4.1
-	github.com/golang/protobuf v1.3.1 // indirect
 	github.com/gorilla/mux v1.7.0
+	github.com/kr/pretty v0.1.0 // indirect
 	github.com/mailru/easyjson v0.0.0-20190312143242-1de009706dbe // indirect
-	github.com/mattn/go-isatty v0.0.7 // indirect
-	github.com/stretchr/objx v0.1.1 // indirect
+	github.com/stretchr/testify v1.3.0 // indirect
 	github.com/swaggo/files v0.0.0-20190110041405-30649e0721f8 // indirect
-	github.com/swaggo/gin-swagger v1.1.0 // indirect
 	github.com/swaggo/http-swagger v0.0.0-20190219123516-d2bf002bf65b
 	github.com/swaggo/swag v1.4.1
-	github.com/ugorji/go/codec v0.0.0-20190316192920-e2bddce071ad // indirect
-	golang.org/x/crypto v0.0.0-20190313024323-a1f597ede03a // indirect
 	golang.org/x/net v0.0.0-20190313220215-9f648a60d977 // indirect
-	golang.org/x/sys v0.0.0-20190316082340-a2f829d7f35f // indirect
 	golang.org/x/tools v0.0.0-20190318005222-63e6ed9258fa // indirect
 	google.golang.org/appengine v1.4.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
